Add files via upload
Ok so, my additions broke everything :/ great. Anyways I managed to make it functional again but you have to input the int twice in the terminal for some reason. I think it's a issue with not closing the scanner at the right time. Turns out in my first attempt having multiple scanners is a bad idea....